requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.3
kafka-python==2.0.2
Mako==1.3.5
MarkupSafe==3.0.1
mysqlclient==2.2.4
oauthlib==3.2.0
pyasn1==0.6.1
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
PyJWT==2.3.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2022.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.12.2
uvicorn==0.27.1
strawberry-graphql
